Improve routing and type hinting across application
In this commit, adjustments were made to parts of the Filament resource. The `mount` function and 'after' function in the 'ServerProvisioning.php' file were updated to ensure that upon server creation or deletion, the user is redirected to an appropriate page. In 'ProvisionServer.php', the handle and failed methods were changed to include return type hinting for better code reliability. Uncommented lines in the 'SiteResource.php' file to ensure functioning server breadcrumbs. Heavily modified 'ListSitesServer.php' and 'DatabaseServer.php' to improve management of related records and the creation of databases. Changes were necessary to enhance user experience and strengthen the robustness of the code.