Refactor frequency field state setting in CronServer
Simplified the if-else conditional structure used to set the state of the 'frequency' field in the CronServer.php file. Now, if the record's expression exists in the predefinedFrequencyOptions array of the Cron class, the record's expression is used to set the state. Otherwise, it defaults to 'custom'. This optimizes the code for readability and performance.