"Refine ServerResource and Policies code interface"
The code was adjusted primarily for greater clarity and type safety. In CronServer.php, a recordTitleAttribute was added to the method to display a command title on the server page. Furthermore, the HasFactory trait was removed from the Database class model, and explicit return type void was added to the booted method. DatabasePolicy's viewAny and create methods were also updated to not receive a User as a parameter since it wasn't used. These modifications will make the code interface more readable and enhance type safety.