"Removed unnecessary TailwindCSS configuration and updated dependencies"
In this commit, the redundant TailwindCSS configuration file was removed as it is no longer required due to the shift to Laravel's Vite plugin for managing TailwindCSS. This simplifies the resulting configuration and enhances maintainability. The vite.config.js file was also modified to simplify the plugin's refresh path configuration.

Additionally, this commit includes updates to package dependencies in both the package.json and package-lock.json files. These updates encompass a variety of important patches and vulnerability fixes that improve the overall robustness and security of the application.

Lastly, minor changes have been incorporated into the DatabaseServer.php file to prepare for future enhancements in managing database users.