Added user role selection to DatabaseServer
This commit introduces a new feature where the user can select between different roles ('member', 'administrator', 'owner') for the users part of the server. This was achieved by adding functionalities to the existing Select and Repeater classes in `DatabaseServer.php`.

Another update is the addition of a scaffold for a header action that will display the count of users associated with the server. This function is currently commented out as it will be fully developed in the forthcoming commits.

This will improve user management and provide a better overview of server users.