Update DatabaseServer functionality
Enabled the user selection feature in the DatabaseServer. The 'manage users' action count and 'existing users' selection are now operational. Two new components (Repeater and Select) from Filament\Forms were added to support this feature. The changes allow database users linked with a server to be properly managed and selected from the UI. Selection implementation was initially commented out for further investigation, but has now been successfully integrated.

Additionally, an uncommented "expandableLimitedList" code line is now marked as "to be checked". Fixing this function call is beyond this