Refactor event broadcasting in DaemonUpdated and DaemonDeleted
Removed dependencies on ProtoneMedia\Splade package in the DaemonUpdated and DaemonDeleted events and replaced it with Laravel's standard broadcasting mechanism. Simplified the usage of Channel and cleanup of code referring to old package. Refactored the broadcastOn() and broadcastWith() methods to use native Laravel broadcasting features. This results in less dependencies and more standard, cleaner code.