Update JobOnServerFailed to modify error display URL
The 'toMail()' and 'via()' methods in the `app/Notifications/JobOnServerFailed.php` file have been modified to remove the `$notifiable` argument. Additionally, ServerResource has been imported and used to alter the URL in the 'action' line of the notification. This was done to change the way server daemon installation failure errors are displayed to the user, giving us more flexibility in managing and fixing these issues.