Moved activity log creation to DaemonServer file
Shifted the creation of activity log entry from InstallDaemon job to the DaemonServer model. This change ensures the consistency of logs as the activity is now logged right after a daemon is created or updated, rather than during the daemon installation process. This also avoids potential situations where installation might fail but the log entry would still be created.